+++ examples/cluster_multi_az/outputs.tf
@@ -1,0 +1,11 @@
+output "cluster_id" {
+  value = osdgoogle_cluster.cluster.id
+}
+
+output "api_url" {
+  value = osdgoogle_cluster.cluster.api_url
+}
+
+output "console_url" {
+  value = osdgoogle_cluster.cluster.console_url
+}